move generating collections to a separate file
